Move `test_infra` under `tools/`
