+++ .github/CODEOWNERS
@@ -142,7 +142,7 @@ test/ttmlir/Dialect/StableHLO @tenstorrent/forge-developers-mlir-stablehlo
 /tools/pykernel/ @xanderchin @vprajapati-tt
 
 # TTIRBuilder
-/python/test_infra @ctodTT @tapspatel
+/tools/test_infra @ctodTT @tapspatel
 
 # tt-explorer
 /tools/explorer/ @tenstorrent/forge-developers-mlir-explorer
