+++ test/python/golden/test_ttir_ops_llmbox.py
@@ -5,8 +5,8 @@ import torch
 import pytest
 
 from typing import List, Tuple
-from ttmlir.test_utils import compile_to_flatbuffer
-from ttmlir.ttir_builder import Operand, TTIRBuilder, Shape
+from test_infra.test_utils import compile_to_flatbuffer
+from test_infra.ttir_builder import Operand, TTIRBuilder, Shape
 
 pytestmark = pytest.mark.llmbox
